Tabs!!!
Code tabs are here! It's now a LOT easier to work on multiple files at
once. I might even dare to say this is starting to look a lot more like
a real IDE. I also added sparse documentation (the function names
usually make it clear what's going on). This will be the last feature
update for a while